requirements.txt
# Core
Flask==3.1.3
waitress==3.0.2
APScheduler==3.11.2
python-dotenv==1.2.2

# Database
SQLAlchemy==2.0.49
PyMySQL==1.1.2

# Indirect (pinned for reproducibility)
blinker==1.9.0
click==8.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.8
typing_extensions==4.15.0
tzlocal==5.3.1
